[MS-0] Added configurable caching
