+++ src/test/resources/application-disabled-cache.properties
@@ -1,0 +1,1 @@
+nutaxi.route.service.cache.enabled=false
