+++ src/test/resources/application-test.properties
@@ -1,2 +1,4 @@
 spring.redis.host=localhost
-spring.redis.port=6378
+spring.redis.port=6379
+
+nutaxi.route.service.cache.enabled=true
