+++ src/test/resources/bootstrap-disabled-cache.properties
@@ -1,0 +1,3 @@
+spring.application.name=nutaxi-route-service
+spring.cloud.config.enabled=false
+eureka.client.enabled=false
